requirements.txt
# Dependências do projeto

# Biblioteca para fazer requisições HTTP (usada para comunicação com a API da OpenAI)
requests

# Biblioteca para carregar variáveis de ambiente a partir do arquivo .env
python-dotenv
